Fix errors and warnings from devtools::check()
-Unify two methods of seasonOmptimCal
-Add the scope of external functions or import them into the NAMESPACE
-Change parameter names to match generic function
-Fix type mismatch in src/cohortModel.c
-Change erroneous function name getParams2diff1 -> setParams2diff1
-Add generics for var and cumsum
-Remove nSteps parameter from combineLH_Env. Use broods from LH instead.

+++ src/cohortModel.c
@@ -36,7 +36,7 @@ int survdist(int n0, double surv, double varSurv, double limit, double ** pSurv)
 {
 #ifdef IEEE_754
     /* NaNs propagated correctly */
-    if (ISNAN(n0) || ISNAN(surv) || ISNAN(limit)) {
+    if (ISNAN((double) n0) || ISNAN(surv) || ISNAN(limit)) {
 //       pSurv = NULL;
         return 0;
     }
@@ -119,7 +119,7 @@ int fertdist(double * years, int maxYear, int broods, int B, double surv, double
 {
 #ifdef IEEE_754
     /* NaNs propagated correctly */
-    if (ISNAN(* years) || ISNAN(maxYear)|| ISNAN(broods) || ISNAN(B) || ISNAN(surv) || ISNAN(varSurv)) {
+    if (ISNAN(* years) || ISNAN((double) maxYear)|| ISNAN((double) broods) || ISNAN((double) B) || ISNAN(surv) || ISNAN(varSurv)) {
 //         pFert = NULL;
         return 0;
     }
